Aplicando Manager genérico de filtro por data em mais pontos do código
